refactor: use tracker::Service in StatisticsImporter
Instead of using the API client directly. TrackerService is easier to
mock becuase you only need to build simple app structs instead of
reqwest responses.
